Enable handling record save response in different formats
