+++ skygear/src/main/java/io/skygear/skygear/MultiRecordSaveResponseHandler.java
@@ -1,0 +1,14 @@
+package io.skygear.skygear;
+
+
+/**
+ * The Response Handler for Saving Multiple Records
+ */
+public abstract class MultiRecordSaveResponseHandler
+        extends RecordSaveResponseBaseHandler<Record[]>
+{
+    @Override
+    protected final void onSaveRecordsSuccess(Record[] records) {
+        this.onSaveSuccess(records);
+    }
+}
